requirements.txt
streamlit
pandas
email-validator
Flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2 
validate_email
pyDNS 